+++ KaguyaProjectV2/KaguyaBot/Core/Commands/Currency/FishGame.cs
@@ -8,8 +8,10 @@ using KaguyaProjectV2.KaguyaBot.Core.KaguyaEmbed;
 using KaguyaProjectV2.KaguyaBot.DataStorage.DbData.Models;
 using KaguyaProjectV2.KaguyaBot.DataStorage.DbData.Queries;
 using System;
+using System.Configuration;
 using System.Linq;
 using System.Threading.Tasks;
+using KaguyaProjectV2.KaguyaBot.Core.Configurations.Models;
 using KaguyaProjectV2.KaguyaBot.Core.Extensions;
 
 namespace KaguyaProjectV2.KaguyaBot.Core.Commands.Currency
@@ -56,18 +58,18 @@ namespace KaguyaProjectV2.KaguyaBot.Core.Commands.Currency
             
             bool premium = await user.IsPremiumAsync();
             bool bassHatBonus = user.HasActiveBonus(Tool.BASSMASTER_HAT);
-            UserTool masterHat = null;
+            UserTool bassMasterHat = null;
             
             if (bassHatBonus)
             {
-                masterHat = inventory.Tools.Where(x => x.Tool == Tool.BASSMASTER_HAT && x.CurrentDurability > 0)
+                bassMasterHat = inventory.Tools.Where(x => x.Tool == Tool.BASSMASTER_HAT && x.CurrentDurability > 0)
                     .OrderByDescending(x => x.Rank).FirstOrDefault();
             }
 
-            if (masterHat != null)
+            if (bassMasterHat != null)
             {
                 // For ranks 2 and beyond, the cooldown is reduced by 1 second for each rank.
-                FISHING_COOLDOWN_BASSMASTER_HAT -= masterHat.Rank - 1;
+                FISHING_COOLDOWN_BASSMASTER_HAT -= bassMasterHat.Rank - 1;
             }
             
             if (user.LastFished >= DateTime.Now.AddSeconds(-FISHING_COOLDOWN).ToOADate() && !premium && !bassHatBonus ||
@@ -98,10 +100,6 @@ namespace KaguyaProjectV2.KaguyaBot.Core.Commands.Currency
             Random r = new Random();
             double roll = r.NextDouble();
             
-            // This rolls from 0.000 -- 1.000. Increased by +0.05 for each hat rank, increasing your chance to keep it
-            // by up to 0.05. By default, hat has a 3% chance to be lost with each fish cast. If you have a 
-            // rank 5 hat though, the chance to lose it is 0.5%.
-            double? bassMasterHatWind = masterHat == null ? null : r.NextDouble() + 0.005 * (masterHat?.Rank - 1);
             //todo: Keep testing and building out the master hat.
             int fishId = r.Next(int.MaxValue);
             int fishExp;
@@ -269,10 +267,56 @@ namespace KaguyaProjectV2.KaguyaBot.Core.Commands.Currency
             };
             // Fish Embed
             await ReplyAsync(embed: embed.Build());
-            
-            //todo: Add logic for bassmaster hat.
+            await HatDurabilityLogic(inventory);
         }
 
+        private async Task HatDurabilityLogic(Inventory userInventory)
+        {
+            UserTool bassmasterHat = userInventory.Tools?.FirstOrDefault(x => x?.Tool == Tool.BASSMASTER_HAT && x.CurrentDurability > 0);
+
+            if (bassmasterHat == null)
+                return;
+            
+            var r = new Random();
+            var durabilityRng = r.NextDouble();
+            var loseChance = r.NextDouble();
+            bool loseDurability = durabilityRng <= .80 - .05 * (bassmasterHat.Rank - 1); // 80% chance to lose 1 durability on a fish cast, -5% chance for each rank.
+            bool lostHat = loseChance <= .03 - .005 * (bassmasterHat.Rank - 1); // 3% chance to lose hat entirely, -0.5% per rank.
+
+            if (loseDurability)
+            {
+                // Takes away one durability from the hat.
+                bassmasterHat.CurrentDurability -= 1;
+            }
+            
+            if (lostHat)
+            {
+                bassmasterHat.CurrentDurability = 0;
+            }
+
+            if (bassmasterHat.CurrentDurability <= 0)
+            {
+                string[] eventText =
+                {
+                    "has fallen into the water and was swallowed by sharks!",
+                    "was swept away by the wind and was lost in the trees.",
+                    "is damaged beyond repair.",
+                    "was stolen by squirrels while you were in the middle of casting!"
+                };
+
+                int index = r.Next(0, eventText.Length);
+                
+                var embed = new KaguyaEmbedBuilder(EmbedColor.YELLOW)
+                {
+                    Description = $"Oh no! Your `Bassmaster's Hat` {eventText[index]}!"
+                };
+
+                await SendEmbedAsync(embed);
+            }
+
+            //todo: Figure out database updating!
+        }
+        
         private FishType GetFishType(double roll)
         {
             if (roll <= 0.0005) // 1 in 2000 chance. o_o